backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.1
pydantic-settings==2.4.0
SQLAlchemy[asyncio]==2.0.35
asyncpg==0.29.0
python-multipart==0.0.9
orjson==3.10.7
passlib[bcrypt]==1.7.4
boto3==1.35.20
pytesseract==0.3.13
Pillow==10.4.0
stripe==11.3.0
prometheus-client==0.20.0
google-auth==2.34.0
PyJWT[crypto]==2.9.0
requests==2.32.3
